pyproject.toml
[project]
name = "learn-gen-ai"
version = "0.1.0"
description = "simple applications using ai tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.12.0",
    "openai-whisper>=20240930",
    "torch>=2.5.1",
    "transformers>=4.48.0",
]
